fix(region): remove waf cache
