+++ pkg/apis/compute/waf_regexsets.go
@@ -18,30 +18,15 @@ import (
 	"yunion.io/x/onecloud/pkg/apis"
 )
 
-const (
-	WAF_REGEX_SET_STATUS_AVAILABLE     = "available"
-	WAF_REGEX_SET_STATUS_DELETING      = "deleting"
-	WAF_REGEX_SET_STATUS_DELETE_FAILED = "delete_failed"
-)
-
 type WafRegexSetDetails struct {
 	apis.StatusInfrasResourceBaseDetails
 	SWafRegexSet
+	ManagedResourceInfo
+	CloudregionResourceInfo
 }
 
 type WafRegexSetListInput struct {
 	apis.StatusInfrasResourceBaseListInput
-}
-
-type WafRegexSetCacheDetails struct {
-	apis.StatusStandaloneResourceDetails
-	ManagedResourceInfo
-	CloudregionResourceInfo
-	SWafRegexSetCache
-}
-
-type WafRegexSetCacheListInput struct {
-	apis.StatusStandaloneResourceListInput
 	apis.ExternalizedResourceBaseListInput
 
 	ManagedResourceListInput
