+++ pkg/compute/models/waf_rule_groups.go
@@ -16,27 +16,28 @@ package models
 
 import (
 	"context"
+	"fmt"
 
 	"yunion.io/x/cloudmux/pkg/cloudprovider"
 	"yunion.io/x/jsonutils"
-	"yunion.io/x/log"
 	"yunion.io/x/pkg/errors"
 	"yunion.io/x/pkg/util/compare"
 	"yunion.io/x/sqlchemy"
 
+	"yunion.io/x/onecloud/pkg/apis"
 	api "yunion.io/x/onecloud/pkg/apis/compute"
-	identity_api "yunion.io/x/onecloud/pkg/apis/identity"
 	"yunion.io/x/onecloud/pkg/cloudcommon/db"
 	"yunion.io/x/onecloud/pkg/cloudcommon/db/lockman"
 	"yunion.io/x/onecloud/pkg/httperrors"
 	"yunion.io/x/onecloud/pkg/mcclient"
 	"yunion.io/x/onecloud/pkg/util/stringutils2"
-	"yunion.io/x/onecloud/pkg/util/yunionmeta"
 )
 
 type SWafRuleGroupManager struct {
 	db.SStatusInfrasResourceBaseManager
 	db.SExternalizedResourceBaseManager
+	SManagedResourceBaseManager
+	SCloudregionResourceBaseManager
 }
 
 var WafRuleGroupManager *SWafRuleGroupManager
@@ -56,12 +57,10 @@ func init() {
 type SWafRuleGroup struct {
 	db.SStatusInfrasResourceBase
 	db.SExternalizedResourceBase
+	SManagedResourceBase
+	SCloudregionResourceBase
 
-	// 支持的WAF类型,仅is_system=true时有效
-	WafType  cloudprovider.TWafType `width:"40" charset:"ascii" list:"domain" nullable:"false"`
-	Provider string                 `width:"20" charset:"ascii" list:"domain" nullable:"false"`
-	CloudEnv string                 `width:"20" charset:"ascii" list:"domain" nullable:"false"`
-	IsSystem bool                   `nullable:"false" default:"false" list:"domain" update:"domain" create:"optional"`
+	WafType cloudprovider.TWafType `width:"40" charset:"ascii" list:"domain" nullable:"false"`
 }
 
 func (manager *SWafRuleGroupManager) FetchCustomizeColumns(
@@ -74,9 +73,13 @@ func (manager *SWafRuleGroupManager) FetchCustomizeColumns(
 ) []api.WafRuleGroupDetails {
 	rows := make([]api.WafRuleGroupDetails, len(objs))
 	siRows := manager.SStatusInfrasResourceBaseManager.FetchCustomizeColumns(ctx, userCred, query, objs, fields, isList)
+	managerRows := manager.SManagedResourceBaseManager.FetchCustomizeColumns(ctx, userCred, query, objs, fields, isList)
+	regionRows := manager.SCloudregionResourceBaseManager.FetchCustomizeColumns(ctx, userCred, query, objs, fields, isList)
 	for i := range rows {
 		rows[i] = api.WafRuleGroupDetails{
 			StatusInfrasResourceBaseDetails: siRows[i],
+			ManagedResourceInfo:             managerRows[i],
+			CloudregionResourceInfo:         regionRows[i],
 		}
 	}
 	return rows
@@ -96,16 +99,19 @@ func (manager *SWafRuleGroupManager) ListItemFilter(
 		return nil, errors.Wrap(err, "SStatusInfrasResourceBaseManager.ListItemFilter")
 	}
 
-	if query.IsSystem != nil {
-		q = q.Equals("is_system", *query.IsSystem)
+	q, err = manager.SExternalizedResourceBaseManager.ListItemFilter(ctx, q, userCred, query.ExternalizedResourceBaseListInput)
+	if err != nil {
+		return nil, errors.Wrap(err, "SExternalizedResourceBaseManager.ListItemFilter")
 	}
 
-	if len(query.Provider) > 0 {
-		q = q.Equals("provider", query.Provider)
+	q, err = manager.SManagedResourceBaseManager.ListItemFilter(ctx, q, userCred, query.ManagedResourceListInput)
+	if err != nil {
+		return nil, errors.Wrap(err, "SManagedResourceBaseManager.ListItemFilter")
 	}
 
-	if len(query.CloudEnv) > 0 {
-		q = q.Equals("cloud_env", query.CloudEnv)
+	q, err = manager.SCloudregionResourceBaseManager.ListItemFilter(ctx, q, userCred, query.RegionalFilterListInput)
+	if err != nil {
+		return nil, errors.Wrap(err, "SCloudregionResourceBaseManager.ListItemFilter")
 	}
 
 	return q, nil
@@ -117,6 +123,16 @@ func (manager *SWafRuleGroupManager) QueryDistinctExtraField(q *sqlchemy.SQuery,
 	if err == nil {
 		return q, nil
 	}
+	q, err = manager.SManagedResourceBaseManager.QueryDistinctExtraField(q, field)
+	if err == nil {
+		return q, nil
+	}
+
+	q, err = manager.SCloudregionResourceBaseManager.QueryDistinctExtraField(q, field)
+	if err == nil {
+		return q, nil
+	}
+
 	return q, httperrors.ErrNotFound
 }
 
@@ -130,6 +146,15 @@ func (manager *SWafRuleGroupManager) OrderByExtraFields(
 	if err != nil {
 		return nil, errors.Wrap(err, "SStatusInfrasResourceBaseManager.OrderByExtraFields")
 	}
+	q, err = manager.SManagedResourceBaseManager.OrderByExtraFields(ctx, q, userCred, query.ManagedResourceListInput)
+	if err != nil {
+		return nil, errors.Wrap(err, "SManagedResourceBaseManager.OrderByExtraFields")
+	}
+	q, err = manager.SCloudregionResourceBaseManager.OrderByExtraFields(ctx, q, userCred, query.RegionalFilterListInput)
+	if err != nil {
+		return nil, errors.Wrap(err, "SCloudregionResourceBaseManager.OrderByExtraFields")
+	}
+
 	return q, nil
 }
 
@@ -142,7 +167,24 @@ func (manager *SWafRuleGroupManager) ListItemExportKeys(ctx context.Context,
 	if err != nil {
 		return nil, errors.Wrap(err, "SStatusInfrasResourceBaseManager.ListItemExportKeys")
 	}
+	if keys.ContainsAny(manager.SCloudregionResourceBaseManager.GetExportKeys()...) {
+		q, err = manager.SCloudregionResourceBaseManager.ListItemExportKeys(ctx, q, userCred, keys)
+		if err != nil {
+			return nil, errors.Wrap(err, "SCloudregionResourceBaseManager.ListItemExportKeys")
+		}
+	}
+	if keys.ContainsAny(manager.SManagedResourceBaseManager.GetExportKeys()...) {
+		q, err = manager.SManagedResourceBaseManager.ListItemExportKeys(ctx, q, userCred, keys)
+		if err != nil {
+			return nil, errors.Wrap(err, "SManagedResourceBaseManager.ListItemExportKeys")
+		}
+	}
+
 	return q, nil
+}
+
+func (self *SWafRuleGroup) syncRemove(ctx context.Context, userCred mcclient.TokenCredential) error {
+	return self.RealDelete(ctx, userCred)
 }
 
 func (self *SWafRuleGroup) Delete(ctx context.Context, userCred mcclient.TokenCredential) error {
@@ -163,185 +205,122 @@ func (self *SWafRuleGroup) RealDelete(ctx context.Context, userCred mcclient.Tok
 	return self.SStatusInfrasResourceBase.Delete(ctx, userCred)
 }
 
-func (manager *SWafRuleGroupManager) GetWafGroups(cloudEnv string) ([]SWafRuleGroup, error) {
-	q := manager.Query().Equals("cloud_env", cloudEnv).IsTrue("is_system")
-	groups := []SWafRuleGroup{}
-	err := db.FetchModelObjects(WafRuleGroupManager, q, &groups)
-	return groups, err
+func (self *SWafRuleGroup) GetIRegion(ctx context.Context) (cloudprovider.ICloudRegion, error) {
+	region, err := self.GetRegion()
+	if err != nil {
+		return nil, errors.Wrapf(err, "GetRegion")
+	}
+	provider, err := self.GetDriver(ctx)
+	if err != nil {
+		return nil, errors.Wrapf(err, "GetDriver")
+	}
+	return provider.GetIRegionById(region.ExternalId)
 }
 
-func (self *SWafRuleGroup) syncWithCloudSku(ctx context.Context, userCred mcclient.TokenCredential, ext sWafGroup) error {
+func (self *SWafRuleGroup) GetICloudWafRuleGroup(ctx context.Context) (cloudprovider.ICloudWafRuleGroup, error) {
+	if len(self.ExternalId) == 0 {
+		return nil, errors.Wrapf(cloudprovider.ErrNotFound, "empty external id")
+	}
+	iRegion, err := self.GetIRegion(ctx)
+	if err != nil {
+		return nil, errors.Wrapf(err, "GetIRegion")
+	}
+	caches, err := iRegion.GetICloudWafRuleGroups()
+	if err != nil {
+		return nil, errors.Wrapf(err, "GetICloudWafRuleGroups")
+	}
+	for i := range caches {
+		if caches[i].GetGlobalId() == self.ExternalId {
+			return caches[i], nil
+		}
+	}
+	return nil, errors.Wrapf(cloudprovider.ErrNotFound, self.ExternalId)
+}
+
+func (self *SWafRuleGroup) syncWithCloudRuleGroup(ctx context.Context, userCred mcclient.TokenCredential, ext cloudprovider.ICloudWafRuleGroup) error {
 	_, err := db.Update(self, func() error {
-		self.Name = ext.Name
-		self.Description = ext.Description
-		self.IsPublic = true
-		self.Status = api.WAF_RULE_GROUP_STATUS_AVAILABLE
+		self.Status = apis.STATUS_AVAILABLE
+		self.Name = ext.GetName()
+		self.WafType = ext.GetWafType()
+		self.Description = ext.GetDesc()
 		return nil
 	})
-	if err != nil {
-		return errors.Wrapf(err, "db.Update")
+	return err
+}
+
+func (self *SCloudregion) GetRuleGroups(managerId string) ([]SWafRuleGroup, error) {
+	q := WafRuleGroupManager.Query().Equals("cloudregion_id", self.Id)
+	if len(managerId) > 0 {
+		q = q.Equals("manager_id", managerId)
 	}
-	result, err := self.SyncManagedWafRules(ctx, userCred, ext.Rules)
+	ret := []SWafRuleGroup{}
+	err := db.FetchModelObjects(WafRuleGroupManager, q, &ret)
 	if err != nil {
-		return errors.Wrapf(err, "SyncManagedWafRules")
+		return nil, errors.Wrapf(err, "db.FetchModelObjects")
 	}
-	log.Debugf("Sync waf group %s rule result: %s", self.Name, result.Result())
-	return nil
+	return ret, nil
 }
 
-func (manager *SWafRuleGroupManager) newFromCloudWafGroup(ctx context.Context, userCred mcclient.TokenCredential, ext sWafGroup) error {
-	group := &ext.SWafRuleGroup
-	group.SetModelManager(manager, group)
-	group.Status = api.WAF_RULE_GROUP_STATUS_AVAILABLE
-	group.IsPublic = true
-	err := WafRuleGroupManager.TableSpec().Insert(ctx, group)
-	if err != nil {
-		return errors.Wrapf(err, "Insert")
-	}
-	result, err := group.SyncManagedWafRules(ctx, userCred, ext.Rules)
-	if err != nil {
-		return errors.Wrapf(err, "SyncManagedWafRules")
-	}
-	log.Debugf("Sync waf group %s rule result: %s", group.Name, result.Result())
-	return nil
-}
-
-type sWafGroup struct {
-	SWafRuleGroup
-	Rules []SWafRule
-}
-
-func (self sWafGroup) GetGlobalId() string {
-	return self.ExternalId
-}
-
-func (self SWafRule) GetGlobalId() string {
-	return self.ExternalId
-}
-
-func (manager *SWafRuleGroupManager) SyncWafGroups(ctx context.Context, userCred mcclient.TokenCredential, cloudEnv string, isStart bool) compare.SyncResult {
-	lockman.LockRawObject(ctx, cloudEnv, manager.Keyword())
-	defer lockman.ReleaseRawObject(ctx, cloudEnv, manager.Keyword())
+func (self *SCloudregion) SyncWafRuleGroups(ctx context.Context, userCred mcclient.TokenCredential, provider *SCloudprovider, exts []cloudprovider.ICloudWafRuleGroup) compare.SyncResult {
+	lockman.LockRawObject(ctx, WafRuleGroupManager.Keyword(), fmt.Sprintf("%s-%s", self.Id, provider.Id))
+	defer lockman.ReleaseRawObject(ctx, WafRuleGroupManager.Keyword(), fmt.Sprintf("%s-%s", self.Id, provider.Id))
 
 	result := compare.SyncResult{}
 
-	meta, err := yunionmeta.FetchYunionmeta(ctx)
-	if err != nil {
-		result.Error(errors.Wrapf(err, "FetchYunionmeta"))
-		return result
-	}
-
-	exts := []sWafGroup{}
-	err = meta.List(WafRuleManager.Keyword(), cloudEnv, &exts)
-	if err != nil {
-		result.Error(errors.Wrapf(err, "List(%s)", cloudEnv))
-		return result
-	}
-	dbGroup, err := manager.GetWafGroups(cloudEnv)
-	if err != nil {
-		result.Error(errors.Wrapf(err, "GetWafGroups"))
-		return result
-	}
-
-	if isStart && len(dbGroup) > 0 {
-		log.Infof("%s waf group already synced, skip...", cloudEnv)
-		return result
-	}
-
-	removed := make([]SWafRuleGroup, 0)
-	commondb := make([]SWafRuleGroup, 0)
-	commonext := make([]sWafGroup, 0)
-	added := make([]sWafGroup, 0)
-
-	err = compare.CompareSets(dbGroup, exts, &removed, &commondb, &commonext, &added)
+	dbRuleGroups, err := self.GetRuleGroups(provider.Id)
 	if err != nil {
 		result.Error(err)
 		return result
 	}
 
-	for i := 0; i < len(removed); i += 1 {
-		err = removed[i].RealDelete(ctx, userCred)
+	removed := make([]SWafRuleGroup, 0)
+	commondb := make([]SWafRuleGroup, 0)
+	commonext := make([]cloudprovider.ICloudWafRuleGroup, 0)
+	added := make([]cloudprovider.ICloudWafRuleGroup, 0)
+	err = compare.CompareSets(dbRuleGroups, exts, &removed, &commondb, &commonext, &added)
+	if err != nil {
+		result.Error(err)
+		return result
+	}
+
+	for i := 0; i < len(removed); i++ {
+		err := removed[i].syncRemove(ctx, userCred)
 		if err != nil {
 			result.DeleteError(err)
 			continue
 		}
 		result.Delete()
 	}
-	for i := 0; i < len(commondb); i += 1 {
-		err = commondb[i].syncWithCloudSku(ctx, userCred, commonext[i])
+
+	for i := 0; i < len(commondb); i++ {
+		err := commondb[i].syncWithCloudRuleGroup(ctx, userCred, commonext[i])
 		if err != nil {
 			result.UpdateError(err)
 			continue
 		}
 		result.Update()
 	}
-	for i := 0; i < len(added); i += 1 {
-		err = manager.newFromCloudWafGroup(ctx, userCred, added[i])
+
+	for i := 0; i < len(added); i++ {
+		err = self.newFromCloudWafRuleGroup(ctx, userCred, provider, added[i])
 		if err != nil {
 			result.AddError(err)
 			continue
 		}
 		result.Add()
 	}
-
 	return result
 }
 
-func fetchCloudEnvs() ([]string, error) {
-	accounts := []SCloudaccount{}
-	q := CloudaccountManager.Query("provider", "access_url").In("provider", CloudproviderManager.GetPublicProviderProvidersQuery()).Distinct()
-	err := q.All(&accounts)
-	if err != nil {
-		return nil, errors.Wrapf(err, "q.All")
-	}
-	ret := []string{}
-	for i := range accounts {
-		ret = append(ret, api.GetCloudEnv(accounts[i].Provider, accounts[i].AccessUrl))
-	}
-	return ret, nil
-}
-
-func SyncWafGroups(ctx context.Context, userCred mcclient.TokenCredential, isStart bool) {
-	err := func() error {
-		cloudEnvs, err := fetchCloudEnvs()
-		if err != nil {
-			return errors.Wrapf(err, "fetchCloudEnvs")
-		}
-
-		if len(cloudEnvs) == 0 {
-			return nil
-		}
-
-		meta, err := yunionmeta.FetchYunionmeta(ctx)
-		if err != nil {
-			return errors.Wrapf(err, "FetchYunionmeta")
-		}
-
-		index, err := meta.Index(WafRuleManager.Keyword())
-		if err != nil {
-			return errors.Wrapf(err, "getWafIndex")
-		}
-
-		for _, cloudEnv := range cloudEnvs {
-			skuMeta := &SWafRuleGroup{}
-			skuMeta.SetModelManager(WafRuleGroupManager, skuMeta)
-			skuMeta.DomainId = identity_api.DEFAULT_DOMAIN_ID
-			skuMeta.Id = cloudEnv
-
-			oldMd5 := db.Metadata.GetStringValue(ctx, skuMeta, db.SKU_METADAT_KEY, userCred)
-			newMd5, ok := index[cloudEnv]
-			if !ok || newMd5 == yunionmeta.EMPTY_MD5 || len(oldMd5) > 0 && newMd5 == oldMd5 {
-				continue
-			}
-
-			db.Metadata.SetValue(ctx, skuMeta, db.SKU_METADAT_KEY, newMd5, userCred)
-
-			result := WafRuleGroupManager.SyncWafGroups(ctx, userCred, cloudEnv, isStart)
-			log.Debugf("sync %s waf group result: %s", cloudEnv, result.Result())
-		}
-		return nil
-	}()
-	if err != nil {
-		log.Errorf("SyncWafGroups: error: %v", err)
-	}
+func (self *SCloudregion) newFromCloudWafRuleGroup(ctx context.Context, userCred mcclient.TokenCredential, provider *SCloudprovider, ext cloudprovider.ICloudWafRuleGroup) error {
+	ret := &SWafRuleGroup{}
+	ret.SetModelManager(WafRuleGroupManager, ret)
+	ret.Name = ext.GetName()
+	ret.CloudregionId = self.Id
+	ret.ManagerId = provider.Id
+	ret.ExternalId = ext.GetGlobalId()
+	ret.Status = apis.STATUS_AVAILABLE
+	ret.WafType = ext.GetWafType()
+	ret.Description = ext.GetDesc()
+	return WafRuleGroupManager.TableSpec().Insert(ctx, ret)
 }
